Create 1160. Find Words That Can Be Formed by Characters
Input: words = ["cat","bt","hat","tree"], chars = "atach"
Output: 6
Explanation: The strings that can be formed are "cat" and "hat" so the answer is 3 + 3 = 6.

Aim is to check is a word in words is made up of characters from chars string. And return the sum of length of those words.